extract getGlobalConfigDir into its own module to avoid circular import
