+++ source/lib/storage/paths.ts
@@ -1,25 +1,19 @@
 import fs from 'node:fs';
-import os from 'node:os';
 import path from 'node:path';
 import {failed, isFail, Result, succeeded} from '../model/result-types.js';
 import {getRepoRootDir, getStateBranchRoot} from '../../git/git-storage.js';
 
+export {
+	GLOBAL_CONFIG_DIR_NAME,
+	getGlobalConfigDir,
+} from './global-config-dir.js';
+
 export const isLocal = process.env['IS_LOCAL'] === 'true';
 
 export const EPIQ_DIR_NAME = '.epiq';
-export const GLOBAL_CONFIG_DIR_NAME = '.epiq-global';
 export const EVENTS_DIR_NAME = 'events';
 export const MEDIA_DIR_NAME = 'media';
 export const PROJECT_FILE_NAME = 'project.json';
-
-/**
- * Where Epiq's global state (user config, worktrees) lives. Overridable via
- * `EPIQ_GLOBAL_DIR` so tests (and other embedders) can isolate it without
- * touching `HOME`, which git/ssh/npm also rely on.
- */
-export const getGlobalConfigDir = (): string =>
-	process.env['EPIQ_GLOBAL_DIR'] ??
-	path.join(os.homedir(), GLOBAL_CONFIG_DIR_NAME);
 
 export const getEpiqDirPath = (root: string): string =>
 	path.join(root, EPIQ_DIR_NAME);
